[build] Add option to limit imported sources
When using webpack 4 you can rely on [tree shaking](https://webpack.js.org/guides/tree-shaking/) to drop unused sources when creating your Avatar component like the example below.

```javascript
import { createAvatarComponent, TwitterSource } from 'react-avatar';

const Avatar = createAvatarComponent({
    sources: [ TwitterSource ]
});
```

Exported sources:
- GravatarSource
- FacebookSource
- TwitterSource
- GoogleSource
- GithubSource
- SkypeSource
- ValueSource
- SrcSource
- IconSource
- RedirectSource

If you are using a version of webpack that does not support tree shaking or are using a different bundler you'll need to import only those files you need.

```javascript
import createAvatarComponent from 'react-avatar/es/avatar';
import TwitterSource from 'react-avatar/es/sources/Twitter';

const Avatar = createAvatarComponent({
    sources: [ TwitterSource ]
});
```

```javascript
const createAvatarComponent = require('react-avatar/lib/avatar').default;
const TwitterSource = require('react-avatar/lib/sources/Twitter').default;

const Avatar = createAvatarComponent({
    sources: [ TwitterSource ]
});
```
